# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpaceX Member Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MemberStatus = typeof MemberStatus[keyof typeof MemberStatus];


export const MemberStatus = {
  ACTIVE: 'ACTIVE',
  PENDING: 'PENDING',
  SUSPENDED: 'SUSPENDED',
} as const;

export type MemberTier = typeof MemberTier[keyof typeof MemberTier];


export const MemberTier = {
  Explorer: 'Explorer',
  Pioneer: 'Pioneer',
  Vanguard: 'Vanguard',
} as const;

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  member: 'member',
  admin: 'admin',
} as const;

export interface Member {
  id: string;
  email: string;
  name: string;
  tier: MemberTier;
  status: MemberStatus;
  role: MemberRole;
  title?: string | null;
  location?: string | null;
  member_since?: string | null;
  created_at: string;
  updated_at: string;
}

export interface MemberListItem {
  id: string;
  email: string;
  name: string;
  tier: MemberTier;
  status: MemberStatus;
  role: MemberRole;
  member_since?: string | null;
}

export interface CreateMemberInput {
  email: string;
  /** @minLength 1 */
  name: string;
  tier?: MemberTier;
  status?: MemberStatus;
  role?: MemberRole;
  title?: string;
  location?: string;
  member_since?: string;
}

export interface UpdateMemberInput {
  email?: string;
  /** @minLength 1 */
  name?: string;
  tier?: MemberTier;
  status?: MemberStatus;
  role?: MemberRole;
  title?: string | null;
  location?: string | null;
  member_since?: string | null;
}

export type MemberStatsTierBreakdown = {
  explorer: number;
  pioneer: number;
  vanguard: number;
};

export interface MemberStats {
  totalMembers: number;
  activeMembers: number;
  pendingMembers: number;
  suspendedMembers: number;
  adminCount: number;
  tierBreakdown: MemberStatsTierBreakdown;
}

export interface MemberResponse {
  ok: boolean;
  data: Member;
}

export interface MemberListResponse {
  ok: boolean;
  data: Member[];
}

export type MemberSearchResponseData = {
  data: MemberListItem[];
  total: number;
};

export interface MemberSearchResponse {
  ok: boolean;
  data: MemberSearchResponseData;
}

export interface MemberStatsResponse {
  ok: boolean;
  data: MemberStats;
}

export type DeletedResponseData = {
  id: string;
};

export interface DeletedResponse {
  ok: boolean;
  data: DeletedResponseData;
}

export type ApiErrorError = {
  code: string;
  message: string;
};

export interface ApiError {
  ok: boolean;
  error: ApiErrorError;
}

/**
 * Validation failed
 */
export type ValidationErrorResponse = ApiError;

/**
 * Resource not found
 */
export type NotFoundResponse = ApiError;

/**
 * Internal server error
 */
export type ServerErrorResponse = ApiError;

export type SearchMembersParams = {
/**
 * Search by name or email (case-insensitive)
 */
query?: string;
/**
 * Filter by status
 */
status?: MemberStatus;
/**
 * Filter by membership tier
 */
tier?: MemberTier;
/**
 * Filter by role
 */
role?: MemberRole;
/**
 * Max results to return (default 50)
 * @minimum 1
 * @maximum 200
 */
limit?: number;
/**
 * Number of results to skip (for pagination)
 * @minimum 0
 */
offset?: number;
};