# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpaceX Member Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ApiErrorError } from './apiErrorError';

export interface ApiError {
  ok: boolean;
  error: ApiErrorError;
}