# lib/api-zod/src/generated/types/deletedResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpaceX Member Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { DeletedResponseData } from './deletedResponseData';

export interface DeletedResponse {
  ok: boolean;
  data: DeletedResponseData;
}