# lib/api-zod/src/generated/types/memberListItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpaceX Member Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberRole } from './memberRole';
import type { MemberStatus } from './memberStatus';
import type { MemberTier } from './memberTier';

export interface MemberListItem {
  id: string;
  email: string;
  name: string;
  tier: MemberTier;
  status: MemberStatus;
  role: MemberRole;
  member_since?: Date | null;
}