# lib/api-zod/src/generated/types/memberResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpaceX Member Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Member } from './member';

export interface MemberResponse {
  ok: boolean;
  data: Member;
}