# lib/api-zod/src/generated/types/memberRole.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpaceX Member Portal API
 * OpenAPI spec version: 0.1.0
 */

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  member: 'member',
  admin: 'admin',
} as const;