# lib/api-zod/src/generated/types/memberSearchResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpaceX Member Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberSearchResponseData } from './memberSearchResponseData';

export interface MemberSearchResponse {
  ok: boolean;
  data: MemberSearchResponseData;
}