# lib/api-zod/src/generated/types/memberStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpaceX Member Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberStatsTierBreakdown } from './memberStatsTierBreakdown';

export interface MemberStats {
  totalMembers: number;
  activeMembers: number;
  pendingMembers: number;
  suspendedMembers: number;
  adminCount: number;
  tierBreakdown: MemberStatsTierBreakdown;
}