# lib/api-zod/src/generated/types/memberStatsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpaceX Member Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberStats } from './memberStats';

export interface MemberStatsResponse {
  ok: boolean;
  data: MemberStats;
}