# lib/api-zod/src/generated/types/memberTier.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpaceX Member Portal API
 * OpenAPI spec version: 0.1.0
 */

export type MemberTier = typeof MemberTier[keyof typeof MemberTier];


export const MemberTier = {
  Explorer: 'Explorer',
  Pioneer: 'Pioneer',
  Vanguard: 'Vanguard',
} as const;